Fix ruff lint errors blocking CI
Import sorting, torch.nn alias style, redundant int() calls, and unused
noqa directives — all pre-existing, unrelated to recent RunPod/serving
work but blocking docker-build/deploy. Suppresses two intentional
patterns (FastAPI's File(...) DI default, broad except in the MLflow
fallback) with scoped noqa comments instead of restructuring them.

+++ app/inference.py
@@ -39,7 +39,7 @@ def preprocess(image: Image.Image, cfg: dict) -> torch.Tensor:
 
     w, h = image.size
     ratio = short / min(w, h)
-    new_w, new_h = int(round(w * ratio)), int(round(h * ratio))
+    new_w, new_h = round(w * ratio), round(h * ratio)
     image = image.resize((new_w, new_h), Image.BILINEAR)
 
     arr = np.asarray(image, dtype=np.float32) / 255.0
